plugins/powershell/requirements.txt
# List third-party dependencies here, separated by newlines. 
# All dependencies must be version-pinned, eg. requests==1.2.0
# See: https://pip.pypa.io/en/stable/user_guide/#requirements-files
pywinrm[kerberos]==0.5.0
requests-kerberos==0.15.0
requests-credssp==2.0.0
parameterized==0.9.0
requests==2.32.3
cryptography==44.0.1
idna==3.10
